Refactor Functions.java to separate geotools dependent functions
