+++ core/src/main/java/org/apache/sedona/core/spatialOperator/RangeQuery.java
@@ -19,7 +19,7 @@
 
 package org.apache.sedona.core.spatialOperator;
 
-import org.apache.sedona.common.Functions;
+import org.apache.sedona.common.FunctionsGeoTools;
 import org.apache.sedona.core.rangeJudgement.RangeFilter;
 import org.apache.sedona.core.rangeJudgement.RangeFilterUsingIndex;
 import org.apache.sedona.core.spatialRDD.SpatialRDD;
@@ -55,7 +55,7 @@ public class RangeQuery
     {
         U queryGeometry = originalQueryGeometry;
         if (spatialRDD.getCRStransformation()) {
-            queryGeometry = (U) Functions.transform(originalQueryGeometry, spatialRDD.getSourceEpsgCode(), spatialRDD.getTargetEpgsgCode());
+            queryGeometry = (U) FunctionsGeoTools.transform(originalQueryGeometry, spatialRDD.getSourceEpsgCode(), spatialRDD.getTargetEpgsgCode());
         }
 
         if (useIndex == true) {
